requirements.txt
absl-py==2.3.0
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
args==0.1.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
braceexpand==0.1.7
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
clint==0.5.1
clip==0.2.0
clip-score==0.2.1
datasets==3.6.0
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
fairscale==0.4.13
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
hf_transfer==0.1.9
hpsv2==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
image-reward==1.5
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
json_repair==0.47.3
jsonlines==4.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.92.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tiktoken==0.9.0
timm==1.0.19
tokenizer==3.4.5
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
triton==3.3.1
trl==0.19.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webdataset==1.0.2
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
